Revert "Avoid crash if imported in a web worker (#29)"
This reverts commit 6878528b26b6807eb30e6f5dc3159c219696956f.
